% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico.R
\name{bubblewordchart}
\alias{bubblewordchart}
\title{bubblewordchart}
\usage{
bubblewordchart(df, palavra, qt1, qt2, cor = c("blue", "red"),
  transparencia = 0.4)
}
\arguments{
\item{df}{data.frame}

\item{palavra}{nome da coluna do data.frame que contem as palavras}

\item{qt1}{nome da coluna do data.frame que contém a frequência da palavra no sentimento 1}

\item{qt2}{nome da coluna do data.frame que contém a frequência da palavra no sentimento 2}

\item{cor}{vetor com nome das cores do gráfico (sentimento 1, sentimento 2)}

\item{transparencia}{transparencia das cores}
}
\description{
bubblewordchart
}

