% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico.R
\name{df_bubble_prop}
\alias{df_bubble_prop}
\title{df_bubble_prop
 
 função que separa os pontos de um círculo de acordo com a proporção
 definida pelo parametro p}
\usage{
df_bubble_prop(r, p, cx = 0, cy = 0, n = 1e+05)
}
\description{
df_bubble_prop
 
 função que separa os pontos de um círculo de acordo com a proporção
 definida pelo parametro p
}

