% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico.R
\name{gerar_centros}
\alias{gerar_centros}
\title{gerar_centros}
\usage{
gerar_centros(palavra, qt1, qt2, tamanho = 1000, espacamento = 1)
}
\description{
função gerada p/ criar os centros das bolhas, de forma que elas não tenham
intersecção e que se posicionem de acordo com a proporção ente qt1 e qt2.
}

