% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico.R
\name{menor_distancia}
\alias{menor_distancia}
\title{menor_distancia}
\usage{
menor_distancia(pontos, ponto, px = 1, py = 3)
}
\description{
dada uma lista de pontos e um ponto, esta função encontra o ponto da lista
que possui a menor distância do ponto dado
}

