% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico.R
\name{pontos_borda}
\alias{pontos_borda}
\title{pontos_borda
 
 função que dado o centro e o raio de um círculo, retorna os pontos que estão
 em sua borda.
 o n define a quantidade de pontos da borda.}
\usage{
pontos_borda(cx, cy, r, n = 1e+05)
}
\description{
pontos_borda
 
 função que dado o centro e o raio de um círculo, retorna os pontos que estão
 em sua borda.
 o n define a quantidade de pontos da borda.
}

