% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico.R
\name{retirar_pontos_circulo}
\alias{retirar_pontos_circulo}
\title{retirar_pontos_circulo}
\usage{
retirar_pontos_circulo(df, cx, cy, r)
}
\description{
função que dada uma lista de pontos e um círculo (definido pelo seu centro
e raio) retira os pontos da lista que estão dentro deste círculo.
}

