% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico.R
\name{transformar_em_df}
\alias{transformar_em_df}
\title{transformar_em_df
 
 função que une vários círculos em um único dataset.
 ela precisa dos raios, proporções e centros dos círculos}
\usage{
transformar_em_df(df)
}
\description{
transformar_em_df
 
 função que une vários círculos em um único dataset.
 ela precisa dos raios, proporções e centros dos círculos
}

